# docs/swagger.yaml
basePath: /
definitions:
...
  models.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total_pages:
        type: integer
    type: object
...
  models.Profile:
    properties:
      avatar:
        type: string
      email:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
      verified:
        type: boolean
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        example: Request processed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.TransactionHistory:
    properties:
      counterparty_img:
        type: string
      counterparty_name:
        type: string
      counterparty_phone:
        type: string
      counterparty_type:
        type: string
      created_at:
        type: string
      direction:
        type: string
      id:
        type: integer
      total:
        type: number
      type:
        type: string
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      verified:
        type: boolean
    type: object
  models.WeeklySummary:
    properties:
      total_expense:
        example: 250000
        type: integer
      total_income:
        example: 320000
        type: integer
      week_end:
        example: "2025-09-29"
        type: string
      week_start:
        example: "2025-09-23"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API of Prospera wallet systeme
  title: PROSPERA BACKEND
  version: "1.0"
paths:
...
  /user:
    get:
      description: Retrieve the profile information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Profile User
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Profile'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
  /user/:
    patch:
      consumes:
      - multipart/form-data
      description: Update user's fullname, phone, and avatar image
      parameters:
      - description: Full name of the user
        in: formData
        name: fullname
        type: string
      - description: Phone number of the user
        in: formData
        name: phone
        type: string
      - description: Avatar image file
        in: formData
        name: img
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /user/{id}:
    get:
      description: Retrieve a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get User Profile By ID
      tags:
      - User
  /user/all:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: User's list
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
  /user/avatar:
    delete:
      description: Removes the current avatar of the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User's avatar removed
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Remove user avatar
      tags:
      - User
  /user/history:
    get:
      description: Returns paginated list of user's transaction history
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TransactionHistory'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get user transaction history
      tags:
      - User
  /user/history/{id}:
    delete:
      description: Marks a transaction as deleted for the authenticated user (soft
        delete)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request, invalid transaction id
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized, invalid or missing token
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Soft delete a transaction from user history
      tags:
      - User
  /user/password:
    post:
      consumes:
      - application/json
      description: Allows authenticated user to change their password by providing
        the old password and new password
      parameters:
      - description: Password change payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "201":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request, validation error or old password mismatch
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized, invalid or missing token
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - User
  /user/summary:
    get:
      description: Get income and expense summary (daily or weekly)
      parameters:
      - description: daily or weekly
        enum:
        - daily
        - weekly
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.WeeklySummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get user's financial summary
      tags:
      - User
  /user/wallet:
    get:
      description: Fetch the current wallet balance for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Wallet not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"